Add more control items to indicate which parts of test_run to do. (#51)

+++ test_run.py
@@ -34,29 +34,32 @@ subfield_max = 5    # Total of 3x2 sub-fields  (2 per config file)
 data_dir = 'great3_fit_data'    # This should be set up as a sim-link to your Dropbox folder.
 ps_dir = '../inputs/ps/tables' 
 seed = 12345                    # Whatever.  (But not zero.)
-do_config = True    # Do config-related steps?  For now, this is False, because they are not set up
-                    # properly for variable PSF branches.
-
-# Clean up possible residue from previous runs.
-shutil.rmtree(root, ignore_errors=True)
+do_catalogs = True  # Remake the catalogs?
+do_images = True    # Make the images in great3 builder code?
+do_config = True    # Do config-related steps?  
+do_final = True     # Do final packaging steps?
 
 # Build catalogs, etc.
-t1 = time.time()
-# Reduce number of galaxies so it won't take so long
-great3sims.constants.nrows = 10
-great3sims.constants.ncols = 10
-# Reduce number of stars so it won't take so long
-great3sims.constants.min_star_density = 0.01 # per arcmin^2
-great3sims.constants.max_star_density = 0.03
-great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
-               experiments=experiments, obs_type=obs_type, shear_type=shear_type,
-               gal_dir=data_dir, ps_dir=ps_dir,
-               seed=seed, steps=['metaparameters', 'catalogs']
-)
-t2 = time.time()
-print
-print 'Time for great3sims.run up to catalogs = ',t2-t1
-print
+if do_catalogs:
+    # Clean up possible residue from previous runs.
+    shutil.rmtree(root, ignore_errors=True)
+
+    t1 = time.time()
+    # Reduce number of galaxies so it won't take so long
+    great3sims.constants.nrows = 10
+    great3sims.constants.ncols = 10
+    # Reduce number of stars so it won't take so long
+    great3sims.constants.min_star_density = 0.01 # per arcmin^2
+    great3sims.constants.max_star_density = 0.03
+    great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
+                experiments=experiments, obs_type=obs_type, shear_type=shear_type,
+                gal_dir=data_dir, ps_dir=ps_dir,
+                seed=seed, steps=['metaparameters', 'catalogs']
+    )
+    t2 = time.time()
+    print
+    print 'Time for great3sims.run up to catalogs = ',t2-t1
+    print
 
 # Make list of config file names and directories:
 dirs = []
@@ -94,7 +97,7 @@ if do_config:
             name, ext = os.path.splitext(config_name)
             new_name = '%s_%02d.yaml'%(name,i)
             print config_name,'->',new_name
-            new_config_names.append(new_name)
+            #new_config_names.append(new_name)
             shutil.copy(os.path.join(root,config_name),os.path.join(root,new_name))
         for config_name in psf_config_names:
             name, ext = os.path.splitext(config_name)
@@ -129,33 +132,36 @@ if do_config:
     print 'Total time for galsim = ',t2-t1
     print
 
-    # Move these files to a different name, so we can compare to the ones built in the next step.
-    for ind in range(len(dirs)):
-        dir = dirs[ind]
-        n_epoch = n_epochs[ind]
-        for i in range(subfield_min, subfield_max+1):
-            for j in range(0, n_epoch):
-                f1 = os.path.join(dir,'image-%03d-%1d.fits'%(i,j))
-                f2 = os.path.join(dir,'yaml_image-%03d-%1d.fits'%(i,j))
-                shutil.move(f1,f2)
-                f1 = os.path.join(dir,'starfield_image-%03d-%1d.fits'%(i,j))
-                f2 = os.path.join(dir,'yaml_starfield_image-%03d-%1d.fits'%(i,j))
-                shutil.move(f1,f2)
+    if do_images:
+        # Move these files to a different name, so we can compare to the ones built in the next 
+        # step.
+        for ind in range(len(dirs)):
+            dir = dirs[ind]
+            n_epoch = n_epochs[ind]
+            for i in range(subfield_min, subfield_max+1):
+                for j in range(0, n_epoch):
+                    f1 = os.path.join(dir,'image-%03d-%1d.fits'%(i,j))
+                    f2 = os.path.join(dir,'yaml_image-%03d-%1d.fits'%(i,j))
+                    shutil.move(f1,f2)
+                    f1 = os.path.join(dir,'starfield_image-%03d-%1d.fits'%(i,j))
+                    f2 = os.path.join(dir,'yaml_starfield_image-%03d-%1d.fits'%(i,j))
+                    shutil.move(f1,f2)
 
 # Build images using great3sims.run
-t1 = time.time()
-great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
-               experiments=experiments, obs_type=obs_type, shear_type=shear_type,
-               gal_dir=data_dir, ps_dir=ps_dir,
-               seed=seed, steps=['gal_images', 'psf_images']
-)
-t2 = time.time()
-print
-print 'Time for great3sims.run images = ',t2-t1
-print
+if do_images:
+    t1 = time.time()
+    great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
+                experiments=experiments, obs_type=obs_type, shear_type=shear_type,
+                gal_dir=data_dir, ps_dir=ps_dir,
+                seed=seed, steps=['gal_images', 'psf_images']
+    )
+    t2 = time.time()
+    print
+    print 'Time for great3sims.run images = ',t2-t1
+    print
 
 # Check that images are the same.
-if do_config:
+if do_images and do_config:
     print 'Checking diffs: (No output means success)'
     sys.stdout.flush()
     for ind in range(len(dirs)):
@@ -174,26 +180,27 @@ if do_config:
     print 'End diffs.'
     print
 
-# Measure star parameters required for metric
-t1 = time.time()
-great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
-               experiments=experiments, obs_type=obs_type, shear_type=shear_type,
-               gal_dir=data_dir, ps_dir=ps_dir,
-               seed=seed, steps=['star_params']
-)
-t2 = time.time()
-print
-print 'Time for great3sims.run star_params = ',t2-t1
-print
+if do_final:
+    # Measure star parameters required for metric
+    t1 = time.time()
+    great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
+                experiments=experiments, obs_type=obs_type, shear_type=shear_type,
+                gal_dir=data_dir, ps_dir=ps_dir,
+                seed=seed, steps=['star_params']
+    )
+    t2 = time.time()
+    print
+    print 'Time for great3sims.run star_params = ',t2-t1
+    print
 
 
-# Now package up the data that should be public, and truth tables
-t1 = time.time()
-great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
-               experiments=experiments, obs_type=obs_type, shear_type=shear_type,
-               gal_dir=data_dir, seed=seed, steps=['packages']
-)
-t2 = time.time()
-print
-print 'Time for great3sims.run packages = ',t2-t1
-print
+    # Now package up the data that should be public, and truth tables
+    t1 = time.time()
+    great3sims.run(root, subfield_min=subfield_min, subfield_max=subfield_max,
+                experiments=experiments, obs_type=obs_type, shear_type=shear_type,
+                gal_dir=data_dir, seed=seed, steps=['packages']
+    )
+    t2 = time.time()
+    print
+    print 'Time for great3sims.run packages = ',t2-t1
+    print
